separate libjulia-internal and libjulia-codegen
makes codegen optional via a plugin interface

Co-authored-by: Julian Samaroo <jpsamaroo@jpsamaroo.me>

Add special `@` character to `LOADER_BUILD_DEP_LIBS` and friends

This allows us to mark a library as `"special"` so that the loader
doesn't attempt to open it blindly, but rather interprets it as a list
of positional arguments, that it knows how to interpret.  We strictly
require exactly the number of special libraries, to help avoid errors in
the future as we add to this list.

Strip windows runtime symbols from the windows import library

Without stripping these symbols out of the import library, we end up with duplicate symbol definition errors.

Co-authored-by: Elliot Saba <staticfloat@gmail.com>